+++ deepstruct/scalable.py
@@ -11,7 +11,7 @@ from deepstruct.sparse import MaskedDeepFFN
 from deepstruct.util import kullback_leibler
 
 
-class ScalableFamily(object):
+class ScalableDeepFFN(object):
     _precision: int = 2
     _epsilon: float = 0.001
     _entropy_equivalence_epsilon: float = 0.1
@@ -56,13 +56,13 @@ class ScalableFamily(object):
         return MaskedDeepFFN(input_shape, output_shape, layers)
 
     def __eq__(self, other):
-        if not isinstance(other, ScalableFamily):
+        if not isinstance(other, ScalableDeepFFN):
             return False
 
         return self.entropy_similarity(other) < self.entropy_similarity_epsilon
 
     def entropy_similarity(self, other) -> float:
-        assert isinstance(other, ScalableFamily)
+        assert isinstance(other, ScalableDeepFFN)
         max_length = np.maximum(len(self._proportions), len(other._proportions))
         return kullback_leibler(
             np.pad(self._proportions, (0, max_length - len(self._proportions))),
@@ -74,6 +74,9 @@ class ScalableFamily(object):
 
 
 class ScalableDAN(object):
+    _cached_scaled_structure: LayeredGraph = None
+    _vertex_correspondences: dict = None
+
     _structure: LayeredGraph
     _proportion_map: dict
 
@@ -107,64 +110,83 @@ class ScalableDAN(object):
     def grow(self, scale: int) -> LayeredGraph:
         assert scale > 0
 
-        graph_scaled = CachedLayeredGraph()
-
-        num_nodes = 0
-        vertex_correspondences = {}
         for layer in self.structure.layers:
             for v in self.structure.get_vertices(layer):
-                size = int(np.round(self.proportions[v] * scale))
-                vertex_correspondences[v] = np.arange(num_nodes, num_nodes + size)
-                graph_scaled.add_nodes_from(
-                    vertex_correspondences[v]
-                )  # ['v%s_%s' % (v, idx) for idx in range(size)])
-                num_nodes += size
+                self.scale(v, int(np.round(self.proportions[v] * scale)))
 
-                for (source_vertex, _) in self.structure.in_edges(v):
-                    # print(source_vertex, _)
-                    # print('nodes source:', vertex_correspondences[source_vertex])
-                    # print('nodes target:', vertex_correspondences[v])
-                    # print(list(itertools.product(vertex_correspondences[source_vertex], vertex_correspondences[v])))
-                    graph_scaled.add_edges_from(
-                        itertools.product(
-                            vertex_correspondences[source_vertex],
-                            vertex_correspondences[v],
-                        )
-                    )
+        return self._cached_scaled_structure
+
+    def scale(self, vertex, size: int):
+        graph_scaled = (
+            self._cached_scaled_structure
+            if self._cached_scaled_structure is not None
+            else CachedLayeredGraph()
+        )
+        vertex_correspondences = (
+            self._vertex_correspondences
+            if self._vertex_correspondences is not None
+            else {}
+        )
+        nodes_offset = len(graph_scaled.nodes)
+
+        if vertex not in vertex_correspondences:
+            vertex_correspondences[vertex] = np.array([])
+
+        vertex_correspondences[vertex] = np.concatenate(
+            [
+                vertex_correspondences[vertex],
+                np.arange(nodes_offset, nodes_offset + size),
+            ]
+        )
+        graph_scaled.add_nodes_from(
+            vertex_correspondences[vertex]
+        )  # ['v%s_%s' % (v, idx) for idx in range(size)])
+        nodes_offset += size
+
+        for (source_vertex, _) in self.structure.in_edges(vertex):
+            graph_scaled.add_edges_from(
+                itertools.product(
+                    vertex_correspondences[source_vertex],
+                    vertex_correspondences[vertex],
+                )
+            )
+
+        self._cached_scaled_structure = graph_scaled
+        self._vertex_correspondences = vertex_correspondences
 
         return graph_scaled
 
     def build(self, input_shape, output_shape, scale: int) -> nn.Module:
         assert scale > 0
         graph_scaled = self.grow(scale)
+        self._cached_scaled_structure = None
+        self._vertex_correspondences = None
         return MaskedDeepDAN(input_shape, output_shape, graph_scaled)
 
 
 if __name__ == "__main__":
-    graph = CachedLayeredGraph()
-    graph.add_edge(0, 3)
-    graph.add_edge(0, 5)
-    graph.add_edge(1, 3)
-    graph.add_edge(1, 4)
-    graph.add_edge(1, 5)
-    graph.add_edge(2, 3)
-    graph.add_edge(2, 4)
+    g1 = CachedLayeredGraph()
+    g1.add_edge(0, 3)
+    g1.add_edge(0, 5)
+    g1.add_edge(1, 3)
+    g1.add_edge(1, 4)
+    g1.add_edge(1, 5)
+    g1.add_edge(2, 3)
+    g1.add_edge(2, 4)
 
-    graph.add_edge(3, 5)
-    graph.add_edge(3, 6)
-    graph.add_edge(4, 5)
-    graph.add_edge(4, 7)
+    g1.add_edge(3, 5)
+    g1.add_edge(3, 6)
+    g1.add_edge(4, 5)
+    g1.add_edge(4, 7)
 
-    graph.add_edge(5, 7)
+    g1.add_edge(5, 7)
 
     props = {
         v: p
-        for v, p in zip(
-            graph.nodes, np.random.dirichlet(np.ones(len(graph.nodes)) * 100)
-        )
+        for v, p in zip(g1.nodes, np.random.dirichlet(np.ones(len(g1.nodes)) * 100))
     }
     print(props)
-    fam = ScalableDAN(graph, props)
+    fam = ScalableDAN(g1, props)
     model = fam.build(8, 4, 1000)
     print(model)
     for lay in model.layers_main_hidden: